Xilinx: ARM: DT: Add PS_CLK frequency to ep107
This patch provides a PS_CLK frequency through the DT
for the ep107 platform. The provided frequency is far too
high and may brake a lot of subsystems. But on the other hand,
it results in a working UART, which enables to boot into a
somewhat usable system.

Signed-off-by: Soren Brinkmann <soren.brinkmann@xilinx.com>
(cherry picked from commit 37db21864faf89318e0362a1196097233bc44c97)