Xilinx: ARM: DMA: Updated to latest kernel changes.
In the latest kernel, there are some API changes and also
there is a new enum defined. This patch is created to
update these changes. The changes are in brief,
1. API 'prep_slave_sg' is modified to take an additional
   argument 'void *context'.
2. A new enum 'dma_transfer_direction' has been added.
   This replaces 'dma_data_direction' enum.
All the files which need this changes are updated.

Signed-off-by: SrikanthT <sthokal@xilinx.com>
(cherry picked from commit ccafdb536fdb4e4eb670a41b0eed5ba28b414499)