Fix IRQs-off-tracer for x86_64
Earlier kernels abused the frame pointer to store the
old stack pointer in SAVE_ARGS_IRQ. This behaviour was
changed in upstream commit a2bbe75089d, which breaks
the IRQs-off tracing code for x86_64.

Fix this by basing the argument calculations for
ipipe_trace_begin and ipipe_trace_end on the base
pointer stored in %rsi.

Signed-off-by: Gilles Chanteperdrix <gilles.chanteperdrix@xenomai.org>
Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>
Tested-by: Gernot Hillier <gernot.hillier@siemens.com>