x86/ipipe: disable stack overflow check
Testing for linux task stack overflow can't be done reliably when the
pipeline is active, due to the potential deferral of the timer
handler. In such a case, the stack pointer value from the register set
of the interrupted task could be spuriously checked against the next
incoming task's stack.

We might partially re-enable this feature later, by determining
whether the timer handler was indeed called on behalf of the
interrupted stack.