Arm:Zynq: Clear reboot status register 0x0F000000 bit
FSBL check the reboot status register 0x0F000000 bits to decide
which the partition it should work on. If the bitstream is presented in
the partition, it programs the bistream.

This commit is a temporary workaround to reload the bitstream after
linux reset.

Signed-off-by: Jason Wu <jason.wu@petalogix.com>
Signed-off-by: Michal Simek <monstr@monstr.eu>
(cherry picked from commit 0b64df09a9033887b75a3cdcc52b12d4baaae58a)