arm: zynq: Add boot up frequency OPP
We cannot know the CPU frequency the system boots up with at
compile time.
So, we uncoditionally add this frequency to the OPP table at
runtime.
Since this might result in duplicates in the frequency table,
other points are added only when it is not present in the OPP
list already.

Signed-off-by: Soren Brinkmann <soren.brinkmann@xilinx.com>
Acked-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit cd32815f6ca1aead12f984d8ebced96bb0294297)