arm: zynq: Clean up OPP initialization
Instead of using a fixed set of frequncies for the OPP table, the
frequency operating points are added as following:
1. boot up CPU frequency (F_boot)
2. F_boot / 2
3. F_boot / 3
4. Minium frequency (200 MHz)
The actual frequencies may vary since the HW may not able to generate the exact
frequency.
If any of the frequencies from 2 - 4 fall below the minimum it is not available.

Signed-off-by: Soren Brinkmann <soren.brinkmann@xilinx.com>
Acked-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit 909c493dfb4623656f94a0edd412b5c5615bdda5)