net: ethernet: emacps: Do no use emac power management
This is workaround for problem with busybox udhcp client
which is not able to obtain IP address when PM is on.
But on the other hand we need PM in menuconfig for
others driver that's why I am blocking PM code in the driver.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit 012ad7b10bf82b80775f8a0075933d360593160d)