ipipe: Add root domain guard to preempt_schedule
When we start allowing preempt_disable/enable over non-root domains,
there is the risk to re-enable preemption in non-root contexts. If also
TIF_NEED_RESCHED is set, this will result in a call to preempt_schedule.
Therefore we have to catch any rescheduling over non-root domains on
preempt_schedule entry.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>