zynq: Fixed phy address in QEMU device tree
The QEMU xilinx implementation only has a phy at address 23, isntead of 7 as listed in
the device tree. The of_mdiobus_register function will fail due to not being able to
read the phy ID

Note: This could also be worked around via commit 6bd47ac2e434611e52027155438d7b4ad3c76bdb
which adds a fallback method where the device is assumed to exist because it is listed in the
device tree.