powerpc/ipipe: MPC5200: Fix problem with external interrupts
The MPC5200 external interrupt are not working at all in the Linux
context with the current I-pipe support enabled. For example a
PCI firewire chip connected to an external interrupt doesn't generate
interrupts any more using the Linux FW drivers. With I-pipe
disabled or without the current I-pipe patch for the mpc52xx_pic.c
file, the external interrupts work just fine.

This patch now introduces the xxx_mask_ack() function for these
external interrupts as already done for the other interrupts handled
in this driver. With this change, the interrupts work regardless
if I-pipe support is enabled or not.

Additionally a small coding style issue has been included (spaces
around "-").

Signed-off-by: Stefan Roese <sr@denx.de>
Cc: Philippe Gerum <rpm@xenomai.org>