ipipe: Register function tracer for direct and exclusive invocation
The current I-pipe tracer is best called directly from mcount. This
reduces the overhead and avoids troubles with intermediate trampoline
functions /wrt locking.

We can revise this once our RT domains are compatible with preemption
and RCU based locking, or we are even able to replace the I-pipe tracer
with ftrace.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>