ipipe: Retry clocksource validation in watchdog handler before failing
The realtime domain may interrupt the clocksource watchdog right between
its reading of the tested clock and the reference value. We do not want
to make this sequence atomic as some clocksources may have high access
latencies, but we also do not want to simply drop the watchdog because
broken hardware could otherwise remain unnoticed.

The solution is to read the watchdog clocksource twice, before and after
the tested clock, and retry this until the watchdog clock does not
progress excessively between the measurement points.

The watchdog is generic, but it currently only monitors x86's TSC.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>