dma: xilinx: Ensure segments sg list is ordered
When terminating the dma engine, its possible that the free segment
software list could be built in a way that no longer matches the
hardware ordering of the segments. If this occurs, dma transfer
corruption could occur, due to a mismatch between the descriptors
fetched by hardware and those configured by software.

To essure the sg hardware chain ordering matches the ordering of the
free segment list, rebuild the list based on the hardware ordering when
calling the dma_terminate function.

Signed-off-by: Matt Fornero <matt.fornero@mathworks.com>